Use exhaustive pose3 alignment
We are using exhaustive alignment which loops over all pairs of poses.
We use the pose AUC metric to pick the best alignment among all candidates.
