+++ gtsfm/multi_view_optimizer.py
@@ -2,6 +2,7 @@
 
 Authors: Ayush Baid, John Lambert
 """
+
 from typing import Dict, List, Optional, Tuple
 
 import dask
@@ -13,6 +14,7 @@ from gtsam import Pose3, Rot3, Unit3
 
 import gtsfm.common.types as gtsfm_types
 import gtsfm.utils.graph as graph_utils
+import gtsfm.utils.alignment as alignment_utils
 from gtsfm.averaging.rotation.rotation_averaging_base import RotationAveragingBase
 from gtsfm.averaging.translation.translation_averaging_base import TranslationAveragingBase
 from gtsfm.bundle.global_ba import GlobalBundleAdjustment
@@ -165,7 +167,7 @@ class MultiViewOptimizer:
         ]
 
         # Align the sparse multi-view estimate before BA to the ground truth pose graph.
-        ba_input_graph = dask.delayed(ba_input_graph.align_via_Sim3_to_poses)(gt_wTi_list)
+        ba_input_graph = dask.delayed(alignment_utils.align_gtsfm_data_via_Sim3_to_poses)(ba_input_graph, gt_wTi_list)
 
         return ba_input_graph, ba_result_graph, viewgraph_two_view_reports_graph, multiview_optimizer_metrics_graph
