+++ gtsfm/utils/geometry_comparisons.py
@@ -2,152 +2,21 @@
 
 Authors: Ayush Baid, John Lambert
 """
-import logging
-from typing import List, Optional, Tuple
 
-import gtsam
+from typing import List, Optional
+
 import numpy as np
-from gtsam import Point3, Pose3, Pose3Pairs, Rot3, Rot3Vector, Similarity3, Unit3
+from gtsam import Point3, Pose3, Rot3, Unit3
 from scipy.spatial.transform import Rotation
+
+import gtsfm.utils.logger as logger_utils
+import gtsfm.utils.alignment as alignment_utils
 
 EPSILON = np.finfo(float).eps
 
-logger = logging.getLogger(__name__)
+logger = logger_utils.get_logger()
 
-
-def align_rotations(aRi_list: List[Optional[Rot3]], bRi_list: List[Optional[Rot3]]) -> List[Rot3]:
-    """Aligns the list of rotations to the reference list by using Karcher mean.
-
-    Args:
-        aRi_list: Reference rotations in frame "a" which are the targets for alignment
-        bRi_list: Input rotations which need to be aligned to frame "a"
-
-    Returns:
-        aRi_list_: Transformed input rotations previously "bRi_list" but now which
-            have the same origin as reference (now living in "a" frame)
-    """
-    aRb_list = [
-        aRi.compose(bRi.inverse()) for aRi, bRi in zip(aRi_list, bRi_list) if aRi is not None and bRi is not None
-    ]
-    if len(aRb_list) > 0:
-        aRb = gtsam.FindKarcherMean(Rot3Vector(aRb_list))
-    else:
-        aRb = Rot3()
-
-    # Apply the coordinate shift to all entries in input.
-    return [aRb.compose(bRi) if bRi is not None else None for bRi in bRi_list]
-
-
-def align_poses_sim3_ignore_missing(
-    aTi_list: List[Optional[Pose3]], bTi_list: List[Optional[Pose3]]
-) -> Tuple[List[Optional[Pose3]], Similarity3]:
-    """Align by similarity transformation, but allow missing estimated poses in the input.
-
-    Note: this is a wrapper for align_poses_sim3() that allows for missing poses/dropped cameras.
-    This is necessary, as align_poses_sim3() requires a valid pose for every input pair.
-
-    We force SIM(3) alignment rather than SE(3) alignment.
-    We assume the two trajectories are of the exact same length.
-
-    Args:
-        aTi_list: reference poses in frame "a" which are the targets for alignment
-        bTi_list: input poses which need to be aligned to frame "a"
-
-    Returns:
-        aTi_list_: transformed input poses previously "bTi_list" but now which
-            have the same origin and scale as reference (now living in "a" frame)
-        aSb: Similarity(3) object that aligns the two pose graphs.
-    """
-    assert len(aTi_list) == len(bTi_list)
-
-    # only choose target poses for which there is a corresponding estimated pose
-    corresponding_aTi_list = []
-    valid_camera_idxs = []
-    valid_bTi_list = []
-    for i, bTi in enumerate(bTi_list):
-        if bTi is not None:
-            valid_camera_idxs.append(i)
-            valid_bTi_list.append(bTi)
-            corresponding_aTi_list.append(aTi_list[i])
-
-    valid_aTi_list_, aSb = align_poses_sim3(aTi_list=corresponding_aTi_list, bTi_list=valid_bTi_list)
-
-    num_cameras = len(aTi_list)
-    # now at valid indices
-    aTi_list_ = [None] * num_cameras
-    for i in range(num_cameras):
-        if i in valid_camera_idxs:
-            aTi_list_[i] = valid_aTi_list_.pop(0)
-
-    return aTi_list_, aSb
-
-
-def align_poses_sim3(aTi_list: List[Pose3], bTi_list: List[Pose3]) -> Tuple[List[Pose3], Similarity3]:
-    """Align two pose graphs via similarity transformation. Note: poses cannot be missing/invalid.
-
-    We force Sim(3) alignment rather than SE(3) alignment.
-    We assume the two trajectories are of the exact same length.
-
-    Args:
-        aTi_list: reference poses in frame "a" which are the targets for alignment
-        bTi_list: input poses which need to be aligned to frame "a"
-
-    Returns:
-        aTi_list_: transformed input poses previously "bTi_list" but now which
-            have the same origin and scale as reference (now living in "a" frame)
-        aSb: Similarity(3) object that aligns the two pose graphs.
-    """
-    assert len(aTi_list) == len(bTi_list)
-
-    valid_pose_tuples = [
-        pose_tuple
-        for pose_tuple in list(zip(aTi_list, bTi_list))
-        if pose_tuple[0] is not None and pose_tuple[1] is not None
-    ]
-    n_to_align = len(valid_pose_tuples)
-    if n_to_align < 2:
-        logger.error("SIM(3) alignment uses at least 2 frames; Skipping")
-        return bTi_list, Similarity3(Rot3(), np.zeros((3,)), 1.0)
-
-    ab_pairs = Pose3Pairs(valid_pose_tuples)
-
-    aSb = Similarity3.Align(ab_pairs)
-
-    if np.isnan(aSb.scale()) or aSb.scale() == 0:
-        # we have run into a case where points have no translation between them (i.e. panorama).
-        # We will first align the rotations and then align the translation by using centroids.
-        # TODO: handle it in GTSAM
-
-        # Align the rotations first, so that we can find the translation between the two panoramas.
-        aSb = Similarity3(aSb.rotation(), np.zeros((3,)), 1.0)
-        aTi_list_rot_aligned = [aSb.transformFrom(bTi) for _, bTi in valid_pose_tuples]
-
-        # Fit a single translation motion to the centroid.
-        aTi_centroid = np.array([aTi.translation() for aTi, _ in valid_pose_tuples]).mean(axis=0)
-        aTi_rot_aligned_centroid = np.array([aTi.translation() for aTi in aTi_list_rot_aligned]).mean(axis=0)
-
-        # Construct the final Sim(3) transform.
-        aSb = Similarity3(aSb.rotation(), aTi_centroid - aTi_rot_aligned_centroid, 1.0)
-
-    aSb = Similarity3(R=aSb.rotation(), t=aSb.translation(), s=aSb.scale())
-
-    # Provide a summary of the estimated alignment transform.
-    aRb = aSb.rotation().matrix()
-    atb = aSb.translation()
-    rz, ry, rx = Rotation.from_matrix(aRb).as_euler("zyx", degrees=True)
-    logger.debug("Sim(3) Rotation `aRb`: rz=%.2f deg., ry=%.2f deg., rx=%.2f deg.", rz, ry, rx)
-    logger.debug(f"Sim(3) Translation `atb`: [tx,ty,tz]={str(np.round(atb,2))}")
-    logger.debug("Sim(3) Scale `asb`: %.2f", float(aSb.scale()))
-
-    aTi_list_ = []
-    for bTi in bTi_list:
-        if bTi is None:
-            aTi_list_.append(None)
-        else:
-            aTi_list_.append(aSb.transformFrom(bTi))
-
-    logger.debug("Pose graph Sim(3) alignment complete.")
-    return aTi_list_, aSb
+MAX_ALIGNMENT_ERROR = np.finfo(np.float32).max
 
 
 def compare_rotations(
@@ -182,7 +51,7 @@ def compare_rotations(
     bRi_list = [bRi_list[i] for i in bRi_valid]
 
     # frame 'a' is the target/reference, and bRi_list will be transformed
-    aRi_list_ = align_rotations(aRi_list, bRi_list)
+    aRi_list_ = alignment_utils.align_rotations(aRi_list, bRi_list)
     relative_rotations_angles = np.array(
         [compute_relative_rotation_angle(aRi, aRi_) for (aRi, aRi_) in zip(aRi_list, aRi_list_)], dtype=np.float32
     )
